+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added new DNS resolver client
+- Added new HTTP Client
+
 ## [1.10.1] - 2024-07-04
 
 ### Fixed